start new concurrency model
