+++ packages/core/src/test/audio.test.tsx
@@ -1,43 +1,46 @@
 import {render} from '@testing-library/react';
 import type {ComponentType} from 'react';
-import React, { useContext} from 'react';
+import React, {useContext} from 'react';
 import {Audio} from '../audio';
 import {Internals} from '../internals';
+import {TimelineRoot} from '../TimelineRoot';
 
 const Wrapper: React.FC<{
 	children: React.ReactNode;
 }> = ({children}) => {
 	const compositions = useContext(Internals.CompositionManager);
 	return (
-		<Internals.RemotionRoot>
-			<Internals.CompositionManager.Provider
-				// eslint-disable-next-line react/jsx-no-constructed-context-values
-				value={{
-					...compositions,
-					compositions: [
-						{
-							height: 1080,
-							width: 1080,
-							fps: 30,
-							durationInFrames: 30,
-							id: 'markup',
-							nonce: 0,
-							component: React.lazy(() =>
-								Promise.resolve({
-									default: (() => null) as ComponentType<unknown>,
-								})
-							),
-							defaultProps: undefined,
-							folderName: null,
-							parentFolderName: null,
-						},
-					],
-					currentComposition: 'markup',
-				}}
-			>
-				{children}
-			</Internals.CompositionManager.Provider>
-		</Internals.RemotionRoot>
+		<TimelineRoot>
+			<Internals.RemotionRoot>
+				<Internals.CompositionManager.Provider
+					// eslint-disable-next-line react/jsx-no-constructed-context-values
+					value={{
+						...compositions,
+						compositions: [
+							{
+								height: 1080,
+								width: 1080,
+								fps: 30,
+								durationInFrames: 30,
+								id: 'markup',
+								nonce: 0,
+								component: React.lazy(() =>
+									Promise.resolve({
+										default: (() => null) as ComponentType<unknown>,
+									})
+								),
+								defaultProps: undefined,
+								folderName: null,
+								parentFolderName: null,
+							},
+						],
+						currentComposition: 'markup',
+					}}
+				>
+					{children}
+				</Internals.CompositionManager.Provider>
+			</Internals.RemotionRoot>
+		</TimelineRoot>
 	);
 };
