+++ packages/core/src/test/composition-rules.test.tsx
@@ -2,6 +2,7 @@ import {render} from '@testing-library/react';
 import React from 'react';
 import {Composition} from '..';
 import {RemotionRoot} from '../RemotionRoot';
+import {TimelineRoot} from '../TimelineRoot';
 import {expectToThrow} from './expect-to-throw';
 
 const AnyComp: React.FC = () => null;
@@ -45,24 +46,26 @@ describe('Render composition-rules should throw with invalid props', () => {
 		expectToThrow(
 			() =>
 				render(
-					<RemotionRoot>
-						<Composition
-							lazyComponent={() => Promise.resolve({default: AnyComp})}
-							durationInFrames={100}
-							fps={30}
-							height={100}
-							width={100}
-							id="id"
-						/>
-						<Composition
-							lazyComponent={() => Promise.resolve({default: AnyComp})}
-							durationInFrames={100}
-							fps={30}
-							height={100}
-							width={100}
-							id="id"
-						/>
-					</RemotionRoot>
+					<TimelineRoot>
+						<RemotionRoot>
+							<Composition
+								lazyComponent={() => Promise.resolve({default: AnyComp})}
+								durationInFrames={100}
+								fps={30}
+								height={100}
+								width={100}
+								id="id"
+							/>
+							<Composition
+								lazyComponent={() => Promise.resolve({default: AnyComp})}
+								durationInFrames={100}
+								fps={30}
+								height={100}
+								width={100}
+								id="id"
+							/>
+						</RemotionRoot>
+					</TimelineRoot>
 				),
 			/Multiple composition with id id/
 		);
