+++ packages/core/src/test/loop-validation.test.tsx
@@ -1,44 +1,48 @@
 import {render} from '@testing-library/react';
 import type {ComponentType} from 'react';
-import React, { useContext} from 'react';
-import {Internals} from '..';
+import React, {useContext} from 'react';
+import {CompositionManager} from '../CompositionManager';
 import {Loop} from '../loop';
+import {RemotionRoot} from '../RemotionRoot';
+import {TimelineRoot} from '../TimelineRoot';
 import {expectToThrow} from './expect-to-throw';
 
 const Wrapper: React.FC<{
 	children: React.ReactNode;
 }> = ({children}) => {
-	const compositions = useContext(Internals.CompositionManager);
+	const compositions = useContext(CompositionManager);
 	return (
-		<Internals.RemotionRoot>
-			<Internals.CompositionManager.Provider
-				// eslint-disable-next-line react/jsx-no-constructed-context-values
-				value={{
-					...compositions,
-					compositions: [
-						{
-							height: 1080,
-							width: 1080,
-							fps: 30,
-							durationInFrames: 30,
-							id: 'markup',
-							nonce: 0,
-							component: React.lazy(() =>
-								Promise.resolve({
-									default: (() => null) as ComponentType<unknown>,
-								})
-							),
-							defaultProps: undefined,
-							folderName: null,
-							parentFolderName: null,
-						},
-					],
-					currentComposition: 'markup',
-				}}
-			>
-				{children}
-			</Internals.CompositionManager.Provider>
-		</Internals.RemotionRoot>
+		<TimelineRoot>
+			<RemotionRoot>
+				<CompositionManager.Provider
+					// eslint-disable-next-line react/jsx-no-constructed-context-values
+					value={{
+						...compositions,
+						compositions: [
+							{
+								height: 1080,
+								width: 1080,
+								fps: 30,
+								durationInFrames: 30,
+								id: 'markup',
+								nonce: 0,
+								component: React.lazy(() =>
+									Promise.resolve({
+										default: (() => null) as ComponentType<unknown>,
+									})
+								),
+								defaultProps: undefined,
+								folderName: null,
+								parentFolderName: null,
+							},
+						],
+						currentComposition: 'markup',
+					}}
+				>
+					{children}
+				</CompositionManager.Provider>
+			</RemotionRoot>
+		</TimelineRoot>
 	);
 };
